Move password hashing to separate service MODUSERS-319
